moved data seeding into its own file
